verify_card/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-iot
google-cloud-firestore

